add message coverage test
